Introduce the option to select asymmetric io_uring reactor backend
The current condition for the availability of asymmetric_io_uring is the same as for normal (symmetric) io_uring. This is to be further discussed, e.g. whether we want to ensure a sufficient number of CPUs are available for it to be efficient.
Update the docs to include asymmetric_io_uring backend in the list of available backends.

Introduce detection of the availability of asymmetric io_uring reactor backend.

Currently, conditions for the availability of asymmetric io_uring is the same as that of the normal (symmetric) io_uring. It is currently treated as the most preferred option.

Co-authored-by: MrD4rkne <marcinsz.pv@gmail.com>
Co-authored-by: mikolajbilski <mjbilski@gmail.com>
Co-authored-by: pixelkubek
Co-authored-by: witek-formanski <witekformanski@gmail.com>