Introduce the option to select asymmetric io_uring reactor backend
The current condition for the availability of assymetric_io_uring is the same as for normal (symmetric) io_uring. This is to be further discussed, e.g. whether we want to ensure a sufficient number of CPUs are available for it to be efficient.
Update the docs to include asymmetric_io_uring backend in the list of available backends.