apps: rpc_tester: Add STREAM_BIDIRECTIONAL job
This commit extends the rpc_tester with rpc_streaming job that uses
rpc::sink<> and rpc::source<> to stream data between the client and the
server.

We introduce verb STREAM_BIDIRECTIONAL:

- Streaming is bidirectional - the client streams the payload_t to the
  server, the server streams to the client the number of received bytes
on each read portion of data.

- The client sends the configured number of bytes to the server throught
  rpc::sink, while simultaneously reading from the rpc::source received
from the server.

The server, for each stream, starts a background job to read data from
it. For each payload_t it puts the count of received bytes to the
corresponding rpc::sink and discard the data itself.