Add asymmetric io_uring backend data to reactor_config
A given asymmetric io_uring backend will either be provided with a ready io_uring, or a ring_fd, so it can create a new instance and use IORING_SETUP_ATTACH_WQ to attach it to an existing one.
The variants include std::monostate, to avoid having to pass a value to other backends.
To avoid changing struct size depending on conditional compilation, the io_uring is stored as std::any

Co-authored-by: MrD4rkne <marcinsz.pv@gmail.com>
Co-authored-by: mikolajbilski <mjbilski@gmail.com>
Co-authored-by: pixelkubek <150136485+pixelkubek@users.noreply.github.com>
Co-authored-by: witek-formanski <witekformanski@gmail.com>