Add asymmetric io_uring backend data to reactor_config
A given asymmetric io_uring backend will either be provided with a ready io_uring, or a ring_fd, so it can create a new instance and use IORING_SETUP_ATTACH_WQ to attach it to an existing one.
The variant is wrapped in an optional, to avoid having to pass a value to other backends.