Implement new functions defined in asymmetric_uring_factory
is_master_shard checks whether a given shard should have a uring created in smp::configure, so that other shards can attach to it.
get_uring_group_id returns an id of the shard which will have the uring this shard should attach to. The index starts at 0 and is contiguous.