apps: rpc_tester: Add STREAM_UNIDIRECTIONAL job
We introduce an unidirectional streaming to the rpc_streaming job.

Client streams the configured number of bytes, while server sums the
total data length. After client closes the stream, the server sends back
the total number of received bytes to the corresponding sink. Client
waits for EOF on the stream from the server, then finishes the job.