Implement core choice functions in namespace uring
Functions needed to calculate and allocate async worker cpus as well as choose the correct one for each shard.
is_master_shard checks whether a given shard should have a uring created in smp::configure, so that other shards can attach to it.
get_uring_group_id returns an id of the shard which will have the uring this shard should attach to. The index starts at 0 and is contiguous.