Introduce CLI option to specify async workers cpuset
The new option (async-workers-cpuset) is used to cpecify the CPUs on which workers will be spawned. The recommended way to use this option is to specify it alongside --cpuset. When specified with neither --cpuset nor --smp, it will remove the specified cpus from shard allocation. When --smp or --cpuset are specified, it is possible that worker and app cpusets overlap.

Co-authored-by: MrD4rkne <marcinsz.pv@gmail.com>
Co-authored-by: mikolajbilski <mjbilski@gmail.com>
Co-authored-by: pixelkubek <czyszczon.kuba@gmail.com>
Co-authored-by: witek-formanski <witekformanski@gmail.com>